+++ src/kvp/key.rs
@@ -229,7 +229,7 @@ impl FromStr for KeyPrefix {
 }
 
 impl Deref for KeyPrefix {
-    type Target = String;
+    type Target = str;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -247,7 +247,7 @@ where
     T: AsRef<str>,
 {
     fn eq(&self, other: &T) -> bool {
-        self.as_str() == other.as_ref()
+        self.deref() == other.as_ref()
     }
 }
 
@@ -311,7 +311,7 @@ impl FromStr for KeyName {
 }
 
 impl Deref for KeyName {
-    type Target = String;
+    type Target = str;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -329,14 +329,17 @@ where
     T: AsRef<str>,
 {
     fn eq(&self, other: &T) -> bool {
-        self.as_str() == other.as_ref()
+        self.deref() == other.as_ref()
     }
 }
 
 #[cfg(test)]
 mod test {
-    use super::*;
     use rstest::rstest;
+
+    use crate::kvp::Label;
+
+    use super::*;
 
     #[test]
     fn key_with_prefix() {
@@ -398,5 +401,31 @@ mod test {
     fn invalid_key_name(#[case] input: String, #[case] error: KeyNameError) {
         let err = KeyName::from_str(&input).unwrap_err();
         assert_eq!(err, error);
+    }
+
+    #[rstest]
+    #[case("app.kubernetes.io/name", true)]
+    #[case("name", false)]
+    fn key_prefix_deref(#[case] key: &str, #[case] expected: bool) {
+        let label = Label::try_from((key, "zookeeper")).unwrap();
+
+        let is_valid = label
+            .key()
+            .prefix()
+            .is_some_and(|prefix| *prefix == "app.kubernetes.io");
+
+        assert_eq!(is_valid, expected)
+    }
+
+    #[rstest]
+    #[case("app.kubernetes.io/name", true)]
+    #[case("app.kubernetes.io/foo", false)]
+    #[case("name", true)]
+    #[case("foo", false)]
+    fn key_name_deref(#[case] key: &str, #[case] expected: bool) {
+        let label = Label::try_from((key, "zookeeper")).unwrap();
+        let is_valid = *label.key().name() == "name";
+
+        assert_eq!(is_valid, expected);
     }
 }