+++ CHANGELOG.md
@@ -12,6 +12,12 @@ All notable changes to this project will be documented in this file.
 [#723]: https://github.com/stackabletech/operator-rs/pull/723
 [#724]: https://github.com/stackabletech/operator-rs/pull/724
 
+### Changed
+
+- Change Deref target of `KeyPrefix` and `KeyName` from `String` to `str` ([#725]).
+
+[#725]: https://github.com/stackabletech/operator-rs/pull/725
+
 ## [0.62.0] - 2024-01-19
 
 ### Added